requirements.txt
requests
mitmproxy
paho-mqtt
frida-tools
cryptography
